+++ azurerm/internal/services/apimanagement/api_management_policy_resource.go
@@ -1,0 +1,195 @@
+package apimanagement
+
+import (
+	"fmt"
+	"html"
+	"log"
+	"time"
+
+	"github.com/Azure/azure-sdk-for-go/services/apimanagement/mgmt/2019-12-01/apimanagement"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/apimanagement/parse"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/timeouts"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
+)
+
+func resourceArmApiManagementPolicy() *schema.Resource {
+	return &schema.Resource{
+		Create: resourceArmApiManagementPolicyCreateUpdate,
+		Read:   resourceArmApiManagementPolicyRead,
+		Update: resourceArmApiManagementPolicyCreateUpdate,
+		Delete: resourceArmApiManagementPolicyDelete,
+		Importer: &schema.ResourceImporter{
+			State: schema.ImportStatePassthrough,
+		},
+
+		Timeouts: &schema.ResourceTimeout{
+			Create: schema.DefaultTimeout(30 * time.Minute),
+			Read:   schema.DefaultTimeout(5 * time.Minute),
+			Update: schema.DefaultTimeout(30 * time.Minute),
+			Delete: schema.DefaultTimeout(30 * time.Minute),
+		},
+
+		Schema: map[string]*schema.Schema{
+			"api_management_id": {
+				Type:         schema.TypeString,
+				Required:     true,
+				ForceNew:     true,
+				ValidateFunc: azure.ValidateResourceID,
+			},
+
+			"xml_content": {
+				Type:             schema.TypeString,
+				Optional:         true,
+				Computed:         true,
+				ConflictsWith:    []string{"xml_link"},
+				ExactlyOneOf:     []string{"xml_link", "xml_content"},
+				DiffSuppressFunc: XmlWithDotNetInterpolationsDiffSuppress,
+			},
+
+			"xml_link": {
+				Type:          schema.TypeString,
+				Optional:      true,
+				ConflictsWith: []string{"xml_content"},
+				ExactlyOneOf:  []string{"xml_link", "xml_content"},
+			},
+		},
+	}
+}
+
+func resourceArmApiManagementPolicyCreateUpdate(d *schema.ResourceData, meta interface{}) error {
+	client := meta.(*clients.Client).ApiManagement.PolicyClient
+	ctx, cancel := timeouts.ForCreateUpdate(meta.(*clients.Client).StopContext, d)
+	defer cancel()
+
+	apiManagementID := d.Get("api_management_id").(string)
+	id, err := parse.ApiManagementID(apiManagementID)
+	if err != nil {
+		return err
+	}
+	resourceGroup := id.ResourceGroup
+	serviceName := id.ServiceName
+
+	/*
+		Other resources would have a check for d.IsNewResource() at this location, and would error out using `tf.ImportAsExistsError` if the resource already existed.
+		However, this is a sub-resource, and the API always returns a policy when queried, either a default policy or one configured by the user or by this resource.
+		Instead of the usual check, the resource documentation clearly states that any existing policy will be overwritten if the resource is used.
+	*/
+
+	parameters := apimanagement.PolicyContract{}
+
+	xmlContent := d.Get("xml_content").(string)
+	xmlLink := d.Get("xml_link").(string)
+
+	if xmlLink != "" {
+		parameters.PolicyContractProperties = &apimanagement.PolicyContractProperties{
+			Format: apimanagement.RawxmlLink,
+			Value:  utils.String(xmlLink),
+		}
+	} else if xmlContent != "" {
+		// this is intentionally an else-if since `xml_content` is computed
+
+		// clear out any existing value for xml_link
+		if !d.IsNewResource() {
+			d.Set("xml_link", "")
+		}
+
+		parameters.PolicyContractProperties = &apimanagement.PolicyContractProperties{
+			Format: apimanagement.Rawxml,
+			Value:  utils.String(xmlContent),
+		}
+	}
+
+	if parameters.PolicyContractProperties == nil {
+		return fmt.Errorf("Either `xml_content` or `xml_link` must be set")
+	}
+
+	if _, err := client.CreateOrUpdate(ctx, resourceGroup, serviceName, parameters, ""); err != nil {
+		return fmt.Errorf("creating or updating Policy (Resource Group %q / API Management Service %q): %+v", resourceGroup, serviceName, err)
+	}
+
+	resp, err := client.Get(ctx, resourceGroup, serviceName, apimanagement.PolicyExportFormatXML)
+	if err != nil {
+		return fmt.Errorf("retrieving Policy (Resource Group %q / API Management Service %q): %+v", resourceGroup, serviceName, err)
+	}
+	if resp.ID == nil {
+		return fmt.Errorf("Cannot read ID for Policy (Resource Group %q / API Management Service %q): %+v", resourceGroup, serviceName, err)
+	}
+	d.SetId(*resp.ID)
+
+	return resourceArmApiManagementPolicyRead(d, meta)
+}
+
+func resourceArmApiManagementPolicyRead(d *schema.ResourceData, meta interface{}) error {
+	serviceClient := meta.(*clients.Client).ApiManagement.ServiceClient
+	client := meta.(*clients.Client).ApiManagement.PolicyClient
+	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
+	defer cancel()
+
+	id, err := parse.ApiManagementPolicyID(d.Id())
+	if err != nil {
+		return err
+	}
+	resourceGroup := id.ResourceGroup
+	serviceName := id.ServiceName
+
+	serviceResp, err := serviceClient.Get(ctx, resourceGroup, serviceName)
+	if err != nil {
+		if utils.ResponseWasNotFound(serviceResp.Response) {
+			log.Printf("API Management Service %q was not found in Resource Group %q - removing Policy from state!", serviceName, resourceGroup)
+			d.SetId("")
+			return nil
+		}
+
+		return fmt.Errorf("making Read request on API Management Service %q (Resource Group %q): %+v", serviceName, resourceGroup, err)
+	}
+
+	d.Set("api_management_id", serviceResp.ID)
+
+	resp, err := client.Get(ctx, resourceGroup, serviceName, apimanagement.PolicyExportFormatXML)
+	if err != nil {
+		if utils.ResponseWasNotFound(resp.Response) {
+			log.Printf("[DEBUG] Policy (Resource Group %q / API Management Service %q) was not found - removing from state!", resourceGroup, serviceName)
+			d.SetId("")
+			return nil
+		}
+
+		return fmt.Errorf("making Read request for Policy (Resource Group %q / API Management Service %q): %+v", resourceGroup, serviceName, err)
+	}
+
+	if properties := resp.PolicyContractProperties; properties != nil {
+		policyContent := ""
+		if pc := properties.Value; pc != nil {
+			policyContent = html.UnescapeString(*pc)
+		}
+
+		// when you submit an `xml_link` to the API, the API downloads this link and stores it as `xml_content`
+		// as such there is no way to set `xml_link` and we'll let Terraform handle it
+		d.Set("xml_content", policyContent)
+	}
+
+	return nil
+}
+
+func resourceArmApiManagementPolicyDelete(d *schema.ResourceData, meta interface{}) error {
+	client := meta.(*clients.Client).ApiManagement.PolicyClient
+	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
+	defer cancel()
+
+	id, err := parse.ApiManagementPolicyID(d.Id())
+	if err != nil {
+		return err
+	}
+	resourceGroup := id.ResourceGroup
+	serviceName := id.ServiceName
+
+	if resp, err := client.Delete(ctx, resourceGroup, serviceName, ""); err != nil {
+		if !utils.ResponseWasNotFound(resp) {
+			return fmt.Errorf("deleting Policy (Resource Group %q / API Management Service %q): %+v", resourceGroup, serviceName, err)
+		}
+	}
+
+	return nil
+}