+++ azurerm/internal/services/automation/automation_dsc_configuration_resource.go
@@ -52,7 +52,7 @@ func resourceAutomationDscConfiguration() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: validate.AutomationAccountName(),
+				ValidateFunc: validate.AutomationAccount(),
 			},
 
 			"content_embedded": {