+++ azurerm/internal/services/datalake/data_lake_analytics_account_resource.go
@@ -42,7 +42,7 @@ func resourceArmDataLakeAnalyticsAccount() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: validate.DataLakeAccountName(),
+				ValidateFunc: validate.AccountName(),
 			},
 
 			"location": azure.SchemaLocation(),
@@ -71,7 +71,7 @@ func resourceArmDataLakeAnalyticsAccount() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: validate.DataLakeAccountName(),
+				ValidateFunc: validate.AccountName(),
 			},
 
 			"tags": tags.Schema(),