+++ azurerm/internal/services/datashare/data_share_account_resource.go
@@ -44,7 +44,7 @@ func resourceArmDataShareAccount() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: validate.DataShareAccountName(),
+				ValidateFunc: validate.AccountName(),
 			},
 
 			"resource_group_name": azure.SchemaResourceGroupName(),