+++ azurerm/internal/services/eventgrid/eventgrid_event_subscription_resource.go
@@ -34,12 +34,12 @@ func enpointPropertyNames() []string {
 	}
 }
 
-func resourceArmEventGridEventSubscription() *schema.Resource {
+func resourceEventGridEventSubscription() *schema.Resource {
 	return &schema.Resource{
-		Create: resourceArmEventGridEventSubscriptionCreateUpdate,
-		Read:   resourceArmEventGridEventSubscriptionRead,
-		Update: resourceArmEventGridEventSubscriptionCreateUpdate,
-		Delete: resourceArmEventGridEventSubscriptionDelete,
+		Create: resourceEventGridEventSubscriptionCreateUpdate,
+		Read:   resourceEventGridEventSubscriptionRead,
+		Update: resourceEventGridEventSubscriptionCreateUpdate,
+		Delete: resourceEventGridEventSubscriptionDelete,
 
 		Timeouts: &schema.ResourceTimeout{
 			Create: schema.DefaultTimeout(30 * time.Minute),
@@ -576,7 +576,7 @@ func resourceArmEventGridEventSubscription() *schema.Resource {
 	}
 }
 
-func resourceArmEventGridEventSubscriptionCreateUpdate(d *schema.ResourceData, meta interface{}) error {
+func resourceEventGridEventSubscriptionCreateUpdate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).EventGrid.EventSubscriptionsClient
 	ctx, cancel := timeouts.ForCreateUpdate(meta.(*clients.Client).StopContext, d)
 	defer cancel()
@@ -647,10 +647,10 @@ func resourceArmEventGridEventSubscriptionCreateUpdate(d *schema.ResourceData, m
 
 	d.SetId(*read.ID)
 
-	return resourceArmEventGridEventSubscriptionRead(d, meta)
+	return resourceEventGridEventSubscriptionRead(d, meta)
 }
 
-func resourceArmEventGridEventSubscriptionRead(d *schema.ResourceData, meta interface{}) error {
+func resourceEventGridEventSubscriptionRead(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).EventGrid.EventSubscriptionsClient
 	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
 	defer cancel()
@@ -761,7 +761,7 @@ func resourceArmEventGridEventSubscriptionRead(d *schema.ResourceData, meta inte
 	return nil
 }
 
-func resourceArmEventGridEventSubscriptionDelete(d *schema.ResourceData, meta interface{}) error {
+func resourceEventGridEventSubscriptionDelete(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).EventGrid.EventSubscriptionsClient
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
 	defer cancel()